Add support for 16 bit deep images that is part of PDF 1.5.
Expected Differences:

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4888 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/spdiffx.h
@@ -40,7 +40,7 @@ typedef struct stream_PDiff_state_s {
     int case_index;		/* switch index for case dispatch */
     /* The following are updated dynamically. */
     uint row_left;		/* # of bytes left in row */
-    byte prev[s_PDiff_max_Colors];	/* previous sample */
+    uint prev[s_PDiff_max_Colors];	/* previous sample */
 } stream_PDiff_state;
 
 #define private_st_PDiff_state()	/* in spdiff.c */\
